as-is refactor to allow adding sharded labelstore
