+++ internal/component/prometheus/pipeline_test.go
@@ -174,7 +174,7 @@ func BenchmarkPipelines(b *testing.B) {
 }
 
 func newDefaultPipeline(t testing.TB, logger log.Logger) (storage.Appendable, labelstore.LabelStore, testappender.CollectingAppender) {
-	ls := labelstore.New(logger, promclient.DefaultRegisterer)
+	ls := labelstore.New(logger, promclient.DefaultRegisterer, 1)
 	rwAppendable, rwDestination := newRemoteWriteComponent(t, logger, ls)
 	pipelineAppendable := prometheus.NewFanout([]storage.Appendable{rwAppendable}, "", promclient.DefaultRegisterer, ls)
 	scrapeInterceptor := scrape.NewInterceptor("prometheus.scrape.test", livedebugging.NewLiveDebugging(), pipelineAppendable)
@@ -183,7 +183,7 @@ func newDefaultPipeline(t testing.TB, logger log.Logger) (storage.Appendable, la
 }
 
 func newRelabelPipeline(t testing.TB, logger log.Logger) (storage.Appendable, labelstore.LabelStore, testappender.CollectingAppender) {
-	ls := labelstore.New(logger, promclient.DefaultRegisterer)
+	ls := labelstore.New(logger, promclient.DefaultRegisterer, 1)
 	rwAppendable, rwDestination := newRemoteWriteComponent(t, logger, ls)
 	relabelAppendable := newRelabelComponent(t, logger, []storage.Appendable{rwAppendable}, ls)
 	pipelineAppendable := prometheus.NewFanout([]storage.Appendable{relabelAppendable}, "", promclient.DefaultRegisterer, ls)
