+++ internal/component/prometheus/receive_http/receive_http_test.go
@@ -605,7 +605,7 @@ func testOptions(t *testing.T) component.Options {
 		Logger:     util.TestAlloyLogger(t),
 		Registerer: prometheus.NewRegistry(),
 		GetServiceData: func(name string) (interface{}, error) {
-			return labelstore.New(nil, prometheus.DefaultRegisterer), nil
+			return labelstore.New(nil, prometheus.DefaultRegisterer, 1), nil
 		},
 	}
 }
