+++ internal/runtime/module_eval_test.go
@@ -365,7 +365,7 @@ func testOptions(t *testing.T) runtime.Options {
 			http_service.New(http_service.Options{}),
 			clusterService,
 			otelService,
-			labelstore.New(nil, prometheus.DefaultRegisterer),
+			labelstore.New(nil, prometheus.DefaultRegisterer, 1),
 			remotecfgService,
 		},
 	}
